add label filter to all backlogs
as people are opening prs (and prs are issues in github) they are showing up on the main backlog

I could filter out prs directly but instead Ive chosen to filter on a label. This is so that people can also open actual issues on the module without them showing up. It's also already built in so there's no more code to write.

Reasonable people may differ on this