module tabs variation
feedback on fundamentals was that the prep page was too long and the goals should go elsewhere -- I've made this module tabs view to add in this kind of content. I will also set a short piece of coursework to find, read, and answer some basic questions on this content.