markdownify is sometimes weird
https://github.com/gohugoio/hugo/issues/7372

If the text is just one line, markdownify doesn't give us paragraphs, or a way to tell if this is the case. In the PD blocks we quite often get this.

I am constraining phrasing blocks to a readable line length, so this choice has some text spooling out across the page

Therefore I have begrudgingly wrapped these risk areas in a div and constrained that.

But I don't like it