warn on empty markdown
Addresses #333

I haven't put it as an error (which will fail the build) because the nature of this platform is that we will be pulling in UGC from many places and we won't have control of all the markdown. I don't want to throw errors a contributor cannot fix.

You can't exit a render hook with a warning, so I've placed a check at the start for empty .Destination and printed a warning to the console.

## test

Ive created a broken image in an issue here
https://github.com/CodeYourFuture/Module-Template/issues/6

You can pull it in as a block to see the page build with no image (not even broken in this case, which we can argue about if you like...)

```
[[blocks]]
name="MRE broken image"
src="https://github.com/CodeYourFuture/Module-Template/issues/6"
```