EXAMPLE OF HEADINGS ISSUE
So, for various reasons our toc is created on level 2 headings. Sometimes people erroneously use level 1 on the prep view.

I wrote a render hook to spackle this error but actually I think it's just wrong? and the right thing to do is to just write the correct heading level, which is 2 here, and add this to a FAQ or something for when people hit this issue