FEATURE: stub piscine site (#743)
* stub piscine site

I made this by copying an existing site and junking all the extraneous files.

I think I need to merge this and then I can init and mount this as a module to other sites eg itp

* tweak build settings to always deploy

it keeps cancelling. Netlify help says

If you want to ensure that Netlify does not cancel deploys, you might want to check your build settings. Netlify has a feature that automatically cancels deploys if it doesn't detect changes in the repository. This feature is useful for monorepos where changes might occur in a part of the repository that doesn't affect the site.

However, if you want to disable this feature, you can do so by setting the ignore command in your netlify.toml file to a command that always returns 1. Here's an example:

[build]
  ignore = "exit 1"
This command will always return 1, which means "changes detected" to Netlify, and it will not cancel the deploy.

Please note that this is a workaround and might lead to unnecessary builds if changes are made in the repository that don't affect the site. If you're using a monorepo, it would be better to configure your settings so that Netlify correctly detects changes in the parts of the repository that affect the site.

Sources:

Netlify Documentation: File-based configuration
Netlify Community Post: How can I disable automatic git deploys?