stub new sdc site (#764)
* stub new sdc site

what this is and what this is not

I have copypastad org-cyf-itp
Then I renamed the modules and created a dummy sprint with a template dayplan. Then I copypasted that sprint to make the right length per module.

That's all this is. Once this is merged we will need to register it as a hugo module

* rm erroneous piscine

* Remove more piscine

---------

Co-authored-by: Daniel Wagner-Hall <dawagner@gmail.com>