Add netlify deploy script
We were previously not setting --baseURL which means that deploy
previews have internal links to the prod deploys rather than the deploy
previews.

Instead we introduce a script so that we can edit for all of the sites
without needing to go change several site configs each time we want to
make a change.