Issue cloning always goes via curriculum. (#946)
We don't currently have cloning functionality correctly set up on any of
the other sites.

Rather than needing to set it up properly, and maintain copies of the
netlify functions for each site, let's just use one copy of the netlify
functions on curriculum. and redirect back to the appropriate site after
cloning.

If we choose not to do this, we need to also set up a GitHub OAuth app
per site, because a GitHub OAuth app is only allowed one callback
domain.

It's possible we will need to futz with CORS for this, we'll see.

Once this works, I'll delete the other copies of the netlify function code.