Create a placeholder 404 file with the correct permissions (#1030)
Hugo will correctly replace this with the actual rendered 404 page. But
if this file doesn't already exist with the correct permissions, hugo
will for some reason which we can't debug create a placeholder file
with permissions 0444 instead of 0644, and then won't be able to
overwrite it with the correctly rendered file.