Add brief intro to anonymous and arrow functions (#1059)
We started to use anonymous functions in Jest, and we're going to start
using arrow functions in the next sprint. These aren't scary things, so
we simply explain that they exist and are the same as each other.

We even do so _after_ using them in Jest, rather than before, to
emphasise that they're not scary, and that you probably understood what
they were doing when reading the examples without realising.