TEMP files for objective and task shortcodes (#1190)
this is temporary - we probably want to not just use codefencing and use commented out hugo shortcodes so the task lists still work on github. However we also don't want this sort of glob of code and it makes sense to rationalise  all these various ways of getting objectives in order to make the success page cleaner. But that's a task for a clearer head than I currently have, so here, a quick fix so annotation can proceed without breaking the build