Add guide on presenting, and specifically demoing
We rely on demos in a bunch of places (weekly demos, Piscine demos,
Launch demos, CYF+ demos). This guide attempts to prevent some of the
really common issues we see in these demos.

We can expand this over time.

I've specifically carved out a "Presenting" guide with "Demos" as a
sub-section, because a lot of this advice applies to all presentations.
We can add more presentation types as sub-sections as needed.