Introduce taxonomy equivalent of menu map (#1187)
This produces the same view (sharing code to do so), but allows using a
taxonomy instead of a list of menus. The benefit here is a taxonomy
gives you a place (the taxonomy term page) to define metadata that may
be useful in presenting each term (e.g. attaching a description to each
term).

The equivalences between menu and taxonomy maps are:

| Concept                                                       | Menu concept     | Taxonomy concept      |
|---------------------------------------------------------------|------------------|-----------------------|
| A category (e.g. a column to display in the default map view) | A menu           | A term                |
| A value in the category (e.g. a a card to show in the column) | A page in a menu | A value within a term |

There are some interesting compatibility questions in here which I will
comment on in the PR I'm raising for this.

Note: This currently doesn't actually define any interesting metadata
(other than weight) for each term in the tracks taxonomy - I will send
this through as a follow-up PR, to separate out "the idea of a taxonomy
map" from "the details of this particular taxonomy".