Feature: objectives and tasks oh my (#1218)
* TEMP files

this is temporary - we probably want to not just use codefencing and use commented out hugo shortcodes so the task lists still work on github. However we also don't want this sort of glob of code and it makes sense to rationalise  all these various ways of getting objectives in order to make the success page cleaner. But that's a task for a clearer head than I currently have, so here, a quick fix so annotation can proceed without breaking the build

* extract get list partial

* allow tasks and show example

tasks shows up in the same place on the block, but doesn't come through to success

* simplify parsing and lookup objectives

* oh make these very long and specifcally named

* extract learning objectives from readmes

now for issues...

* weird windows line endings in issues api?

* bump energisers and ops blocks to tasks

they won't show up in success any more

* document tasks

* show objectives on issues

* add issue objectives to single issue

* pop in issues list objectives

* redesign with graphql

1. use graphql instead of rest
2. produce a map of "blocks" and pass into existing system

* some docs

* rename to match blocks naming

* Update common-theme/layouts/partials/strings/extract-github-objectives.html

My formatter keeps doing this and I don't know why

Co-authored-by: Daniel Wagner-Hall <daniel@codeyourfuture.io>

* rm spaces as spaces actually will not work

because my regex is APPALLING

* fix backlog filter to show daniel

* add tasks after objectives

Daniel wants both

---------

Co-authored-by: Daniel Wagner-Hall <daniel@codeyourfuture.io>