Share non-caching config across sites
Previously some sites had config and others didn't, and none of them
disabled getremote caching for production.

This meant that things like PR lists may be stale, subject to hugo's
default getremote caching. It also meant that changing this was a bit
whack-a-mole across the sites.

This changes every site to symlink in the same config (so we can
centrally edit all sites), to disable caching for getremote in
production, and to actually use the production config when deploying.