Soften requirements to apply to be a trainee
We do not require that everyone participating in ITP is eligible to
enroll as a trainee. But we do require that to have completed the
Onboarding module you must have applied.

Frame success as _applying_ to enroll as a trainee, rather than framing
success as contingent on being positively enrolled as a trainee.

We want everyone to apply. It's ok if not everyone is eligible. But
we're trying to anchor people to a path of applying.