1068 Tabs Have Got To Go (#1268)
* add usage example

* storing this

so as to not lose it

* mass regex on tabs

might also go to %%

* update docs

* tabs must have more than one tab

* chomp whitespace on multiple choice

now tabs are renderered differently we could trigger accidental escapes via whitespace -- have  done this on the quiz but will look for more

* chomp whitespace in timer

* renderstring tabs

we will never use tabs in an external context anyway as they need js

* chomp whitespace in shortcodes

reduce chance of hugo wrapping these inline elements with p tags

* really chomp whitespace on time

moving comments seems to be the last thing?

* delete old tab shortcode

* replace markdownify where pos

because they are threatening to deprecate it

* didn't even notice myself doing this

* add === to delimiter

* update tab delimiter

* chomp whitespace without surcease

* final chomps and revert render on local

that was for illustration only!

* that answers that question!

ty daniel!

* OMG PRETTIER WHY ARE YOU LIKE THIS