Rename self-study tracks to guided-study, mark AI as one (#1335)
Realistically AI Essentials isn't a track which leads directly to a job
(and which we only run when we have specific jobs) - it's more of a
"build some skills which may help you in jobs" track. Let's be straight
about this, to avoid diluting what we mean by jobs tracks.

And also pretend React is one for now. Sure, someone could run it as a
track if they wanted.