Simplify How This Works (#1415)
Previously this was a three-module course which confused trainees
because a lot of the content is targeted at volunteers.

This simplifies it. It focuses on how to use the curriculum, and cuts
out many of the volunteer-targeted or class-organisation-targeted
pieces.

Our experience is that writing these things down for volunteers doesn't
make them do them in class - we are trying instead to have that
influence by showing up in person and working with them.

This Self Education module (the same as before, with an added Rhythm of
the Week block) focuses on how trainees should be engaging with the
curriculum and class - let's start here, and add things when they seem
useful / to fill gaps.

The old blocks still exist, and can be easily revived, but this stops
them rendering in the course for now.