Add proof of concept Codewars workshops
Tony has been developing these, here's a proof of concept for embedding
them into the curriculum.

Long-term we may want a custom generator that does closer integration
into the static site generator, but for now iframing things in works
fine.

If we like this, I'll set up a codewars-workshops.codeyourfuture.io or
similar.