Fix multiple choice click capturing (#1560)
Different quizzes in the same page apparently share the same ID space
for labels, even though they're in separate shadow roots.

Fortunately, the `<input>` tags are nested inside the `<label>` tags, so
the `for` attribute isn't necessary, they implicitly bind to their
contained tag.

So removing the `for` attribute here preserves the desired behaviour,
while removing the confusion of which element is being targeted.