Name specific site logos (#1580)
* Name specific site logos

Before c7f101d53e11df76248752d000678d988518eb71 we always showed all
logos. This was problematic if an inheriting site tried to supply logos,
as we would show both.

c7f101d53e11df76248752d000678d988518eb71 tried to fix this by just
showing one, but that breaks the fact that we have two that are shown
differently based on media-query. The SVGs were self-styling with class
attributes.

Instead, we always show exactly two, in a container with the relevant
class attribute. This means the inheriting site must also supply two
logos, and makes that contract explicit.

* Don't mark common-theme as deployable

It's not a useful standalone site.