Clean up Databases track
This makes this a much easier to follow independent chunk of work. It
removes a lot of leftover noise from when this was a module in the old
course.

We will want to give a full review and improvement of this in time, but
for now this makes it a useful stand-alone piece of work to follow.

Major changes:
* Removed all non-database content (e.g. job searching).
* Moved from Module-Databases to Module-SQL (a new repo) which has a
  very small number of issues, because there were previously so many
  unrelated things.
* Removed sprint 4 and squished things together a bit into 3 sprints.
* Removed all day plans. None of them contained any SQL-related content
  other than one workshop which the volunteers were meant to create a
  bunch of content for themselves.
* Make the track actually show up in the tracks list.
* Rename Databases -> SQL.